made big changes in the server folder of backend

+++ Server/controllers/Section.js
@@ -2,18 +2,15 @@ const Section = require("../models/Section");
 const Course = require("../models/Course");
 const SubSection = require("../models/SubSection");
 
-//in course we create many section like chapter_1 , chapter_2 and in section we create many subsection like chapter_1 divided into five topic
-// handler function to create a section
+//in course we create many section like chapter_1 , chapter_2 and in section we create many subsection like chapter_1 divided into five topic...
 const createSection = async (req, res) => {
   try {
-    // data fetch
-    const { sectionName, courseId } = req.body;
-    //data validation
+    const { sectionName, courseId } = req.body; //data fetch
     if (!sectionName || !courseId) {
-      return res.status(400).json({
-        success: false,
-        message: "Missing Properties",
-      });
+      //data validation
+      return res
+        .status(400)
+        .json({ success: false, message: "Missing Properties" });
     }
 
     //create section in DB;
@@ -27,10 +24,9 @@ const createSection = async (req, res) => {
     )
       .populate({ path: "courseContent", populate: { path: "subSection" } })
       .exec();
-    // here we have used populate to replace sections/sub-sections both in the updatedCourseDetails
 
-    //return response
     return res.status(200).json({
+      //return response
       success: true,
       message: "Section created successfully",
       updatedCourseDetails,
@@ -44,35 +40,30 @@ const createSection = async (req, res) => {
   }
 };
 
-// it find section that id is matched with sectionid and in that section {sectionName} is updated;
-// handler function to update a section
 const updateSection = async (req, res) => {
   try {
-    // data input/fetch
-    const { sectionName, sectionId, courseId } = req.body;
-    //data validation
+    const { sectionName, sectionId, courseId } = req.body; //data input
     if (!sectionName || !sectionId) {
+      //data validation
       return res
         .status(400)
         .json({ success: false, message: "Missing Properties" });
     }
 
-    // update data
     const section = await Section.findByIdAndUpdate(
       sectionId,
       { sectionName },
       { new: true }
-    );
-    // await Section.findByIdAndUpdate(sectionId, {sectionName}, {new:true}); we can also write only this, will have no effect on code
+    ); //it find section that id is matched with sectionid and in that section {sectionName} is updated;
+    // await Section.findByIdAndUpdate(sectionId, {sectionName}, {new:true}); agar hm itna bhi likhe to koi effect nhi padega;
+    const course = await Course.findById(courseId)
+      .populate({ path: "courseContent", populate: { path: "subSection" } })
+      .exec();
 
-    // const course = await Course.findById(courseId)
-    //   .populate({ path: "courseContent", populate: { path: "subSection" } })
-    //   .exec();
-
-    //return res
     return res.status(200).json({
+      //return res
       success: true,
-      //   data: course,
+      data: course,
       message: "Section Updated Successfully",
     });
   } catch (error) {
@@ -84,42 +75,37 @@ const updateSection = async (req, res) => {
   }
 };
 
-// handler function to delete a section
+// DELETE a section
 const deleteSection = async (req, res) => {
   try {
-    // get ID - assuming that we are sending ID in params
     const { sectionId, courseId } = req.body;
-    // use findByIdAndDelete
+    await Course.findByIdAndUpdate(courseId, {
+      $pull: { courseContent: sectionId },
+    });
+
+    const section = await Section.findById(sectionId);
+    if (!section) {
+      return res
+        .status(404)
+        .json({ success: false, message: "Section not Found" });
+    }
+
+    //delete sub section
+    await SubSection.deleteMany({ _id: { $in: section.subSection } });
+
     await Section.findByIdAndDelete(sectionId);
 
-    // TODO: Do we need to delete the entry from course schema
+    //find the updated course and return
+    const course = await Course.findById(courseId)
+      .populate({
+        //here there is no use of const course , its only store updated course;
+        path: "courseContent", // if you also write without  "const course = " then it also work;
+        populate: {
+          path: "subSection",
+        },
+      })
+      .exec();
 
-    // await Course.findByIdAndUpdate(courseId, {
-    //   $pull: { courseContent: sectionId },
-    // });
-
-    // const section = await Section.findById(sectionId);
-    // if (!section) {
-    //   return res
-    //     .status(404)
-    //     .json({ success: false, message: "Section not Found" });
-    // }
-
-    // //delete sub section
-    // await SubSection.deleteMany({ _id: { $in: section.subSection } });
-
-    // //find the updated course and return
-    // const course = await Course.findById(courseId)
-    //   .populate({
-    //     //here there is no use of const course , its only store updated course;
-    //     path: "courseContent", // if you also write without  "const course = " then it also work;
-    //     populate: {
-    //       path: "subSection",
-    //     },
-    //   })
-    //   .exec();
-
-    // return response
     res.status(200).json({
       success: true,
       message: "Section deleted",
